move unit tests back into single file, delete unused base class
