--- feature/google/src/test/java/BillingWrapperBC5Test.kt
@@ -1,108 +1,0 @@
-package com.revenuecat.purchases.google
-
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.android.billingclient.api.BillingClient
-import com.android.billingclient.api.PurchaseHistoryResponseListener
-import com.android.billingclient.api.QueryPurchaseHistoryParams
-import com.revenuecat.purchases.utils.mockQueryPurchaseHistory
-import io.mockk.every
-import io.mockk.slot
-import org.assertj.core.api.Assertions.assertThat
-import org.assertj.core.api.AssertionsForClassTypes.fail
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.robolectric.annotation.Config
-import java.util.concurrent.CountDownLatch
-
-@RunWith(AndroidJUnit4::class)
-@Config(manifest = Config.NONE)
-class BillingWrapperBC5Test : BillingWrapperTestBase() {
-
-    @Test
-    fun `queryPurchaseHistoryAsync fails if sent invalid type`() {
-        // TODO pull to common once refactoring queryPurchaseHistoryAsync to take RCProductType
-        billingClientStateListener!!.onBillingSetupFinished(billingClientOKResult)
-
-        mockClient.mockQueryPurchaseHistory(
-            billingClientOKResult,
-            emptyList()
-        )
-        var errorCalled = false
-        wrapper.queryPurchaseHistoryAsync(
-            "notValid",
-            {
-                fail("call should not succeed")
-            },
-            {
-                errorCalled = true
-            }
-        )
-        assertThat(errorCalled).isTrue
-    }
-
-    @Test
-    fun `queryPurchaseHistoryAsync only calls one response when BillingClient responds twice`() {
-        var numCallbacks = 0
-
-        val slot = slot<PurchaseHistoryResponseListener>()
-        every {
-            mockClient.queryPurchaseHistoryAsync(
-                any<QueryPurchaseHistoryParams>(),
-                capture(slot)
-            )
-        } answers {
-            slot.captured.onPurchaseHistoryResponse(billingClientOKResult, null)
-            slot.captured.onPurchaseHistoryResponse(billingClientOKResult, null)
-        }
-
-        wrapper.queryPurchaseHistoryAsync(
-            BillingClient.ProductType.SUBS,
-            {
-                numCallbacks++
-            }, {
-                fail("shouldn't be an error")
-            })
-
-        assertThat(numCallbacks == 1)
-    }
-
-    @Test
-    fun `queryPurchaseHistoryAsync only calls one response when BillingClient responds twice from different threads`() {
-        var numCallbacks = 0
-
-        val slot = slot<PurchaseHistoryResponseListener>()
-        val lock = CountDownLatch(2)
-        every {
-            mockClient.queryPurchaseHistoryAsync(
-                any<QueryPurchaseHistoryParams>(),
-                capture(slot)
-            )
-        } answers {
-            Thread {
-                slot.captured.onPurchaseHistoryResponse(billingClientOKResult, null)
-                lock.countDown()
-            }.start()
-
-            Thread {
-                slot.captured.onPurchaseHistoryResponse(billingClientOKResult, null)
-                lock.countDown()
-            }.start()
-        }
-
-        wrapper.queryPurchaseHistoryAsync(
-            BillingClient.ProductType.SUBS,
-            {
-                // ensuring we don't hit an edge case where numCallbacks doesn't increment before the final assert
-                handler.post {
-                    numCallbacks++
-                }
-            }, {
-                fail("shouldn't be an error")
-            })
-
-        lock.await()
-        assertThat(lock.count).isEqualTo(0)
-
-        assertThat(numCallbacks).isEqualTo(1)
-    }
-}
